note_book_requirements.txt
numpy
pandas
torch
opencv-python
streamlit
pillow
matplotlib
tqdm
transformers
torchvision
torchinfo
